requirements.txt
streamlit
langchain
openai
beautifulsoup4
requests
faiss-cpu
python-dotenv
